@@ -3,227 +3,190 @@
 // Function to determine the orientation based on current and next position.
 int getOrientation(int x1, int y1, int x2, int y2) {
     if (x2 == x1) {
-        return y2 > y1 ? 1 : 3; // North or South
+        return y2 > y1 ? 1 : 3;  // North or South
     } else {
-        return x2 > x1 ? 2 : 0; // East or West
+        return x2 > x1 ? 2 : 0;  // East or West
     }
 }
 
-// Processes the actions of an agent, calculating necessary steps, including orientation changes.
-void processAgentActions(const vector<Point>& points, vector<Step>& steps, int currentOrientation, int agentId) {
+// Processes the actions of an agent, calculating necessary steps, including
+// orientation changes.
+void AgentPathToSteps(const vector<Point>& points, vector<Step>& steps,
+                      int currentOrientation, int agentId) {
     steps.clear();
+    // // Print points
+    // std::cout << "Points" << std::endl;
+    // for (const auto& point : points) {
+    //     std::cout << "(" << point.x << ", " << point.y << "), "
+    //               << "time: " << point.time << ", task_id: " << point.task_id
+    //               << std::endl;
+    // }
     // int currentOrientation = 0; // Assume initial orientation is West.
     double currentTime = 0;
     for (size_t i = 0; i < points.size(); ++i) {
         if (i == 0) {
             // Record the initial position and orientation.
-            steps.push_back({points[i].x, points[i].y, currentOrientation, currentTime});
+            steps.push_back(Step(points[i].x, points[i].y, currentOrientation,
+                                 currentTime, points[i].task_id));
         } else {
             // Determine the necessary orientation for the next move.
-            int neededOrientation = getOrientation(points[i-1].x, points[i-1].y, points[i].x, points[i].y);
+            int neededOrientation = getOrientation(
+                points[i - 1].x, points[i - 1].y, points[i].x, points[i].y);
             if (neededOrientation != currentOrientation) {
-                // Change orientation without advancing time.
-                steps.push_back({points[i-1].x, points[i-1].y, neededOrientation, currentTime});
+                // Change orientation without advancing time by adding an
+                // additional step on the previous timestep with the needed
+                // orientation.
+                steps.push_back(Step(points[i - 1].x, points[i - 1].y,
+                                     neededOrientation, currentTime,
+                                     points[i - 1].task_id));
                 currentOrientation = neededOrientation;
             }
             // Move to the next position, increment time only here.
-            steps.push_back({points[i].x, points[i].y, currentOrientation, ++currentTime});
+            steps.push_back(Step(points[i].x, points[i].y, currentOrientation,
+                                 ++currentTime, points[i].task_id));
         }
     }
 }
 
-// Processes the actions of an agent for the continuous case, using provided timestamps.
-void processAgentActionsContinuous(const vector<Point>& points, vector<Step>& steps, int agentId) {
-    steps.clear();
-    int currentOrientation = 0; // Assume initial orientation is West.
-    double currentTime = 0.0;
-    for (size_t i = 0; i < points.size(); ++i) {
-        if (i == 0) {
-            // Record the initial position and orientation.
-            currentTime = points[i].time;
-            steps.push_back({points[i].x, points[i].y, currentOrientation, currentTime});
-        } else {
-            // Determine the necessary orientation for the next move.
-            int neededOrientation = getOrientation(points[i-1].x, points[i-1].y, points[i].x, points[i].y);
-            if (neededOrientation != currentOrientation) {
-                // Change orientation without advancing time.
-                steps.push_back({points[i-1].x, points[i-1].y, neededOrientation, currentTime});
-                currentOrientation = neededOrientation;
-            }
-            // Move to the next position, use the provided timestamp.
-            currentTime = points[i].time;
-            steps.push_back({points[i].x, points[i].y, currentOrientation, currentTime});
-        }
-    }
-}
+std::vector<std::vector<Action>> StepsToActions(
+    const std::vector<std::vector<Step>>& raw_steps, bool flipped_coord) {
+    // Print steps
+    // std::cout << "Raw Steps" << std::endl;
+    // for (size_t i = 0; i < raw_steps.size(); i++) {
+    //     printf("Step of agent: %lu\n", i);
+    //     for (const auto& tmp_step : raw_steps[i]) {
+    //         std::cout << "(" << tmp_step.x << ", " << tmp_step.y << "), "
+    //                   << tmp_step.orientation << ", " << tmp_step.time
+    //                   << ", task_id: " << tmp_step.task_id << std::endl;
+    //     }
+    //     printf("\n");
+    // }
 
-// Parses a single line of input to extract x, y coordinates and assigns sequential times starting at 1.0.
-vector<Point> parseLine(const string& line) {
-    vector<Point> points;
-    stringstream ss(line);
-    char ignore;
-    int x, y;
-    double time = 1.0;
-
-    ss.ignore(100, ':'); // Ignore characters until ':' which precedes the data.
-
-    while (ss >> ignore >> x >> ignore >> y >> ignore) {
-        points.push_back({x, y, time++});
-        ss >> ignore >> ignore; // Skip "->" characters.
-    }
-
-    return points;
-}
-
-// Parses a single line of input to extract x, y, t coordinates for the continuous case.
-vector<Point> parseLineContinuous(const string& line) {
-    vector<Point> points;
-    stringstream ss(line);
-    char ignore;
-    int x, y;
-    double t;
-
-    ss.ignore(100, ':'); // Ignore characters until ':' which precedes the data.
-    // std::cout << "line:: " << line << std::endl;
-    while (ss >> ignore >> x >> ignore >> y >> ignore >> t >> ignore) {
-        points.push_back({x, y, t});
-        // std::cout << "point: " << x << ", " << y << ", " << t << std::endl;
-        ss >> ignore >> ignore; // Skip "->" characters.
-    }
-
-    return points;
-}
-
-std::vector<std::vector<Action>> processActions(const std::vector<std::vector<Step>>& raw_steps, bool flipped_coord) {
     std::vector<std::vector<Action>> plans;
-    int node_id=0;
+    int node_id = 0;
     for (size_t i = 0; i < raw_steps.size(); ++i) {
         std::vector<Action> processedActions;
-        for (size_t j = 1; j < raw_steps[i].size(); j++) {
-            double prev_step_x, prev_step_y, curr_step_x, curr_step_y;
-            if (flipped_coord) {
-                prev_step_x = raw_steps[i][j-1].y;
-                prev_step_y = raw_steps[i][j-1].x;
-                curr_step_x = raw_steps[i][j].y;
-                curr_step_y = raw_steps[i][j].x;
-            } else {
-                prev_step_x = raw_steps[i][j-1].x;
-                prev_step_y = raw_steps[i][j-1].y;
-                curr_step_x = raw_steps[i][j].x;
-                curr_step_y = raw_steps[i][j].y;
-            }
-            Action processedAction;
-            processedAction.robot_id = (int) i;
-            // @jingtian Note: change action start time, to be consistent with the continuous case
-            processedAction.time = raw_steps[i][j-1].time;
-            processedAction.start.first = prev_step_x;
-            processedAction.start.second = prev_step_y;
-            processedAction.goal.first = curr_step_x;
-            processedAction.goal.second = curr_step_y;
-            processedAction.orientation = raw_steps[i][j].orientation;
-            processedAction.nodeID = node_id;
-            Action accesoray_action = processedAction;
 
-            if (processedAction.start == processedAction.goal &&
-                raw_steps[i][j-1].orientation != raw_steps[i][j].orientation) {
-                processedAction.type = 'T';  // Turn
-            } else if (processedAction.start != processedAction.goal &&
-                        raw_steps[i][j-1].orientation == raw_steps[i][j].orientation) {
-                double mid_x = (prev_step_x + curr_step_x)/2.0;
-                double mid_y = (prev_step_y + curr_step_y)/2.0;
-                processedAction.type = 'M';  // Move
+        if (raw_steps[i].empty()) {
+            // If there are no steps for this agent, skip to the next one.
+        }
+
+        // If the agent has only one step, that is the goal. We need to add a
+        // special goal action.
+        else if (raw_steps[i].size() == 1 && raw_steps[i][0].task_id >= 0) {
+            Action goalAction;
+            goalAction.robot_id = (int)i;
+            goalAction.time = raw_steps[i][0].time;
+            if (flipped_coord) {
+                goalAction.start.first = raw_steps[i][0].y;
+                goalAction.start.second = raw_steps[i][0].x;
+            } else {
+                goalAction.start.first = raw_steps[i][0].x;
+                goalAction.start.second = raw_steps[i][0].y;
+            }
+            goalAction.goal = goalAction.start;  // Goal is the same as start
+            goalAction.orientation = raw_steps[i][0].orientation;
+            goalAction.type = 'S';  // Station
+            goalAction.nodeID = node_id;
+            node_id++;
+            goalAction.task_id = raw_steps[i][0].task_id;
+            processedActions.push_back(goalAction);
+            plans.push_back(processedActions);
+            continue;
+        }
+        // More than 2 steps are available, we need to process them
+        else {
+            for (size_t j = 1; j < raw_steps[i].size(); j++) {
+                double prev_step_x, prev_step_y, curr_step_x, curr_step_y;
+                if (flipped_coord) {
+                    prev_step_x = raw_steps[i][j - 1].y;
+                    prev_step_y = raw_steps[i][j - 1].x;
+                    curr_step_x = raw_steps[i][j].y;
+                    curr_step_y = raw_steps[i][j].x;
+                } else {
+                    prev_step_x = raw_steps[i][j - 1].x;
+                    prev_step_y = raw_steps[i][j - 1].y;
+                    curr_step_x = raw_steps[i][j].x;
+                    curr_step_y = raw_steps[i][j].y;
+                }
+                Action processedAction;
+                processedAction.robot_id = (int)i;
+                // @jingtian Note: change action start time, to be consistent
+                // with the continuous case
+                processedAction.time = raw_steps[i][j - 1].time;
                 processedAction.start.first = prev_step_x;
                 processedAction.start.second = prev_step_y;
-                processedAction.goal.first = mid_x;
-                processedAction.goal.second = mid_y;
-                accesoray_action.type = 'M';
-                accesoray_action.start.first = mid_x;
-                accesoray_action.start.second = mid_y;
-                accesoray_action.goal.first = curr_step_x;
-                accesoray_action.goal.second = curr_step_y;
-                accesoray_action.nodeID = node_id+1;
-            } else if (processedAction.start == processedAction.goal &&
-                       raw_steps[i][j-1].orientation == raw_steps[i][j].orientation) {
-                continue;
-            } else {
-                std::cerr << "Invalid case exiting ..." << std::endl;
-                exit(-1);
-            }
+                processedAction.goal.first = curr_step_x;
+                processedAction.goal.second = curr_step_y;
+                processedAction.orientation = raw_steps[i][j].orientation;
+                processedAction.nodeID = node_id;
+                Action accesoray_action = processedAction;
 
-            if (processedAction.type == 'T' and j == (raw_steps[i].size()-1)) {
-                continue;
-            }
-            processedActions.push_back(processedAction);
-            if (processedAction.type == 'M') {
-                processedActions.push_back(accesoray_action);
+                // Task is considered as complete when the accesoray action is
+                // completed
+                // accesoray_action.task_id = raw_steps[i][j].task_id;
+
+                if (processedAction.start == processedAction.goal &&
+                    raw_steps[i][j - 1].orientation !=
+                        raw_steps[i][j].orientation) {
+                    processedAction.type = 'T';  // Turn
+                } else if (processedAction.start != processedAction.goal &&
+                           raw_steps[i][j - 1].orientation ==
+                               raw_steps[i][j].orientation) {
+                    double mid_x = (prev_step_x + curr_step_x) / 2.0;
+                    double mid_y = (prev_step_y + curr_step_y) / 2.0;
+                    processedAction.type = 'M';  // Move
+                    processedAction.start.first = prev_step_x;
+                    processedAction.start.second = prev_step_y;
+                    processedAction.goal.first = mid_x;
+                    processedAction.goal.second = mid_y;
+                    accesoray_action.type = 'M';
+                    accesoray_action.start.first = mid_x;
+                    accesoray_action.start.second = mid_y;
+                    accesoray_action.goal.first = curr_step_x;
+                    accesoray_action.goal.second = curr_step_y;
+                    accesoray_action.nodeID = node_id + 1;
+                } else if (processedAction.start == processedAction.goal &&
+                           raw_steps[i][j - 1].orientation ==
+                               raw_steps[i][j].orientation) {
+                    continue;
+                } else {
+                    std::cerr << "Invalid case exiting ..." << std::endl;
+                    exit(-1);
+                }
+
+                if (processedAction.type == 'T' and
+                    j == (raw_steps[i].size() - 1)) {
+                    continue;
+                }
+                processedActions.push_back(processedAction);
+                if (processedAction.type == 'M') {
+                    processedActions.push_back(accesoray_action);
+                    node_id++;
+                }
                 node_id++;
+
+                // Add special action for goal arrival
+                if (raw_steps[i][j].task_id >= 0) {
+                    Action goalAction = accesoray_action;
+                    goalAction.type = 'S';  // Station
+                    goalAction.task_id = raw_steps[i][j].task_id;
+                    goalAction.nodeID = node_id;
+                    goalAction.time += 1;
+                    node_id++;
+                }
             }
-            node_id++;
         }
         plans.push_back(processedActions);
     }
+
     return plans;
 }
 
-std::vector<std::vector<Action>> processOneRobustActions(const std::vector<std::vector<Step>>& raw_steps, bool flipped_coord) {
-    std::vector<std::vector<Action>> plans;
-    int node_id=0;
-    for (size_t i = 0; i < raw_steps.size(); ++i) {
-        std::vector<Action> processedActions;
-        for (size_t j = 1; j < raw_steps[i].size(); j++) {
-            double prev_step_x, prev_step_y, curr_step_x, curr_step_y;
-            if (flipped_coord) {
-                prev_step_x = raw_steps[i][j-1].y;
-                prev_step_y = raw_steps[i][j-1].x;
-                curr_step_x = raw_steps[i][j].y;
-                curr_step_y = raw_steps[i][j].x;
-            } else {
-                prev_step_x = raw_steps[i][j-1].x;
-                prev_step_y = raw_steps[i][j-1].y;
-                curr_step_x = raw_steps[i][j].x;
-                curr_step_y = raw_steps[i][j].y;
-            }
-            Action processedAction;
-            processedAction.robot_id = (int) i;
-            // @jingtian Note: change action start time, to be consistent with the continuous case
-            processedAction.time = raw_steps[i][j-1].time;
-            processedAction.start.first = prev_step_x;
-            processedAction.start.second = prev_step_y;
-            processedAction.goal.first = curr_step_x;
-            processedAction.goal.second = curr_step_y;
-            processedAction.orientation = raw_steps[i][j].orientation;
-            processedAction.nodeID = node_id;
-
-            if (processedAction.start == processedAction.goal &&
-                raw_steps[i][j-1].orientation != raw_steps[i][j].orientation) {
-                processedAction.type = 'T';  // Turn
-            } else if (processedAction.start != processedAction.goal &&
-                        raw_steps[i][j-1].orientation == raw_steps[i][j].orientation) {
-                processedAction.type = 'M';  // Move
-            } else if (processedAction.start == processedAction.goal &&
-                       raw_steps[i][j-1].orientation == raw_steps[i][j].orientation) {
-                continue;
-            } else {
-                std::cerr << "Invalid case exiting ..." << std::endl;
-                exit(-1);
-            }
-
-            if (processedAction.type == 'T' and j == (raw_steps[i].size()-1)) {
-                continue;
-            }
-            processedActions.push_back(processedAction);
-            node_id++;
-        }
-        plans.push_back(processedActions);
-    }
-    // showActionsPlan(plans);
-    return plans;
-}
-
-void showStepPoints(std::vector<std::vector<Step>>& raw_plan){
+void showStepPoints(std::vector<std::vector<Step>>& raw_plan) {
     for (size_t i = 0; i < raw_plan.size(); i++) {
         printf("Path of agent: %lu\n", i);
-        for (auto& tmp_point: raw_plan[i]) {
+        for (auto& tmp_point : raw_plan[i]) {
             std::cout << "(" << tmp_point.x << ", " << tmp_point.y << "), "
                       << tmp_point.orientation << ", " << tmp_point.time << ", "
                       << std::endl;
@@ -235,67 +198,15 @@ void showStepPoints(std::vector<std::vector<Step>>& raw_plan){
 void showActionsPlan(std::vector<std::vector<Action>>& plans) {
     for (size_t i = 0; i < plans.size(); i++) {
         printf("Path of agent: %lu\n", i);
-        for (auto &action: plans[i]) {
-            std::cout << "        {" << action.robot_id << ", " << action.time << ", "
-                      << std::fixed << std::setprecision(1) << action.orientation << ", '"
-                      << action.type << "', {" << action.start.first << ", " << action.start.second << "}, {"
-                      << action.goal.first << ", " << action.goal.second << "}, " << action.nodeID << "}," << std::endl;
+        for (auto& action : plans[i]) {
+            std::cout << "        {" << action.robot_id << ", " << action.time
+                      << ", " << std::fixed << std::setprecision(1)
+                      << action.orientation << ", '" << action.type << "', {"
+                      << action.start.first << ", " << action.start.second
+                      << "}, {" << action.goal.first << ", "
+                      << action.goal.second << "}, " << action.nodeID << "},"
+                      << std::endl;
         }
         printf("\n");
     }
-}
-
-bool parseEntirePlan(const std::string& input_file,
-                     std::vector<std::vector<Action>>& plans,
-                     double& raw_cost,
-                     bool flipped_coord,
-                     PlanType file_type){
-    plans.clear();
-    ifstream inFile(input_file);
-    if (!inFile.is_open()) {
-        cerr << "Failed to open the file." << endl;
-        return false;
-    }
-
-    if (file_type == DISCRETE) {
-        std::vector<std::vector<Step>> raw_plan;
-        raw_cost = 0.0;
-        string line;
-        int agentId = 0;
-        while (getline(inFile, line)) {
-            std::vector<Step> tmp_plan;
-            if (!line.empty()) {
-                vector<Point> points = parseLine(line);
-                raw_cost += static_cast<double> (points.size());
-                processAgentActions(points, tmp_plan, 0, agentId++);
-            }
-            raw_plan.push_back(tmp_plan);
-        }
-        inFile.close();
-        plans = processActions(raw_plan, flipped_coord);
-        // showActionsPlan(plans);
-        return true;
-    } else if (file_type == CONTINUOUS) {
-        std::vector<std::vector<Step>> raw_plan;
-        raw_cost = 0.0;
-        string line;
-        int agentId = 0;
-        while (getline(inFile, line)) {
-            std::vector<Step> tmp_plan;
-            if (!line.empty()) {
-                vector<Point> points = parseLineContinuous(line);
-                raw_cost += static_cast<double> (points.size());
-                processAgentActionsContinuous(points, tmp_plan, agentId++);
-            }
-            raw_plan.push_back(tmp_plan);
-        }
-        inFile.close();
-        showStepPoints(raw_plan);
-        plans = processActions(raw_plan, flipped_coord);
-        // showActionsPlan(plans);
-        return true;
-    } else {
-        std::cerr << "Unsupported path format!" << std::endl;
-    }
-    exit(-1);
 }
